:boom: [#1895] -- removed submissions temporary file upload endpoint
This has been replaced a while ago by a formio-specific endpoint.
