+++ src/openapi.yaml
@@ -4126,41 +4126,6 @@ paths:
               $ref: '#/components/headers/X-CSRFToken'
             X-Is-Form-Designer:
               $ref: '#/components/headers/X-Is-Form-Designer'
-  /api/v1/submissions/files/upload:
-    post:
-      operationId: submissions_files_upload_create
-      description: |-
-        Bestandsuploadhandler voor het Form.io bestandsupload opslagtype 'url'.
-
-        Bestandsuploads worden tijdelijke opgeslagen en moeten gekoppeld worden aan een inzending.
-
-        Toegang tot dit endpoint vereist een actieve formulier inzending. Niet gekoppelde bestanden worden automatisch verwijderd na 2 dag(en).
-
-        De maximale toegestane upload-bestandsgrootte is `50,0 MB` voor deze instantie. Merk op dat dit inclusief multipart-metadata en boundaries is. De daadwerkelijke maximale bestandsgrootte is dus iets lager dan deze waarde.
-      summary: Create temporary file upload
-      tags:
-      - submissions
-      requestBody:
-        content:
-          multipart/form-data:
-            schema:
-              $ref: '#/components/schemas/TemporaryFileUpload'
-        required: true
-      deprecated: true
-      responses:
-        '200':
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/TemporaryFileUpload'
-          description: ''
-          headers:
-            X-Session-Expires-In:
-              $ref: '#/components/headers/X-Session-Expires-In'
-            X-CSRFToken:
-              $ref: '#/components/headers/X-CSRFToken'
-            X-Is-Form-Designer:
-              $ref: '#/components/headers/X-Is-Form-Designer'
   /api/v1/translations/formio:
     get:
       operationId: translations_formio_retrieve
