+++ src/openforms/submissions/api/urls.py
@@ -1,10 +1,6 @@
 from django.urls import path
 
-from .views import (
-    DownloadSubmissionReportView,
-    TemporaryFileUploadView,
-    TemporaryFileView,
-)
+from .views import DownloadSubmissionReportView, TemporaryFileView
 
 app_name = "submissions"
 
@@ -13,11 +9,6 @@ urlpatterns = [
         "<int:report_id>/<str:token>/download",
         DownloadSubmissionReportView.as_view(),
         name="download-submission",
-    ),
-    path(
-        "files/upload",
-        TemporaryFileUploadView.as_view(),
-        name="temporary-file-upload",
     ),
     path(
         "files/<uuid:uuid>",
